coderockr/requirements-dev.txt
apturl==0.5.2
asgiref==3.4.1
attrs==22.2.0
autopep8==2.0.2
bcrypt==3.2.0
black==23.1.0
bleach==6.0.0
blinker==1.4
boto3==1.26.118
botocore==1.29.118
Brlapi==0.8.3
Brotli==1.1.0
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
command-not-found==0.3
cryptography==39.0.2
cssselect2==0.7.0
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.12
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.11.2
django-adminactions==2.3.0
django-advanced-filters==2.0.0
django-app-novadata==0.1.1
django-appconf==1.0.5
django-avatar==7.1.1
django-browser-reload==1.7.0
django-colorfield==0.8.0
django-cors-headers==4.0.0
django-crum==0.7.9
django-filter==22.1
django-import-export==3.1.0
django-object-actions==4.2.0
django-quill-editor==0.1.40
django-reorder-admin==0.3.1
django-storages==1.13.2
django-templated-mail==1.1.1
django-weasyprint==2.2.1
django-widget-tweaks==1.4.12
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.2
dnspython==2.3.0
docutils==0.19
docx2pdf==0.1.8
drf-spectacular==0.26.0
drf-spectacular-sidecar==2023.3.1
drf-yasg==1.21.7
duplicity==0.8.21
et-xmlfile==1.1.0
fasteners==0.14.1
filelock==3.6.0
flake8==6.0.0
fonttools==4.49.0
future==0.18.2
html5lib==1.1
httplib2==0.20.2
idna==3.4
importlib-metadata==6.0.0
inflection==0.5.1
jaraco.classes==3.2.3
jeepney==0.8.0
jmespath==1.0.1
jsonschema==4.17.3
jsonschema-specifications==2023.12.1
keyring==23.13.1
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==5.1.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupPy==1.14
MarkupSafe==2.0.1
mccabe==0.7.0
mdurl==0.1.2
monotonic==1.6
more-itertools==9.1.0
mutagen==1.45.1
mypy==1.0.1
mypy-extensions==1.0.0
netifaces==0.11.0
nh3==0.2.15
novadata-utils==0.3.3
numpy==1.26.4
oauthlib==3.2.0
odfpy==1.4.1
olefile==0.46
openpyxl==3.1.1
packaging==23.0
pandas==2.1.4
paramiko==2.9.3
pathspec==0.11.0
pexpect==4.8.0
Pillow==9.4.0
pkginfo==1.9.6
platformdirs==3.1.0
protobuf==3.12.4
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycairo==1.20.1
pycodestyle==2.10.0
pycparser==2.21
pycryptodomex==3.11.0
pycups==2.0.1
pydocstyle==6.3.0
pydyf==0.9.0
pyflakes==3.0.1
Pygments==2.14.0
PyGObject==3.42.1
PyJWT==2.7.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyphen==0.14.0
pyRFC3339==1.1
pyrsistent==0.19.3
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-decouple==3.6
python-docx==0.8.11
python3-openid==3.2.0
pytz==2021.3
pywin32-ctypes==0.2.0
pyxattr==0.7.2
pyxdg==0.27
PyYAML==6.0
readme-renderer==37.3
referencing==0.32.1
reportlab==4.2.0
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==13.3.2
rpds-py==0.16.2
s3transfer==0.6.0
SecretStorage==3.3.3
sentry-sdk==1.37.1
simplejson==3.18.3
six==1.16.0
snowballstemmer==2.2.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.2
static3==0.7.0
systemd-python==234
tablib==3.3.0
tinycss2==1.2.1
tomli==2.0.1
tqdm==4.66.2
twine==4.0.2
typing_extensions==4.5.0
tzdata==2023.3
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.14
usb-creator==0.3.7
virtualenv==20.13.0+ds
wadllib==1.3.6
weasyprint==61.1
webencodings==0.5.1
websockets==9.1
xdg==5
xkit==0.0.0
xlrd==2.0.1
xlwt==1.3.0
yt-dlp==2022.4.8
zipp==3.15.0
zopfli==0.2.3
